Move layer styles & selector into separate files
