+++ src/components/flowchart/styles/flowchart.scss
@@ -1,5 +1,6 @@
 @import './node';
 @import './edges';
+@import './layers';
 @import './tooltip';
 
 .pipeline-flowchart {
@@ -8,26 +9,4 @@
 
 .pipeline-flowchart__graph {
   display: block;
-}
-
-.layer {
-  opacity: 0;
-  transition: opacity ease 0.5s;
-  fill: rgba(white, 0.1);
-
-  &:hover {
-    opacity: 1;
-  }
-}
-
-.layer-names-bg {
-  pointer-events: none;
-  fill: rgba(white, 0.1);
-}
-
-.layer-name {
-  font-weight: bold;
-  font-size: 13px;
-  pointer-events: none;
-  fill: white;
 }
